refactor: validation to own file
