+++ src/components/common/Dialogs/TokenManagementDialog/TokenManagementDialog.tsx
@@ -1,8 +1,6 @@
 import React, { useState } from 'react';
 import { defineMessages } from 'react-intl';
 import { useNavigate } from 'react-router-dom';
-import { string, object, array, boolean, InferType } from 'yup';
-import { isAddress } from 'ethers/lib/utils';
 
 import Dialog, { DialogProps, ActionDialogProps } from '~shared/Dialog';
 import { ActionHookForm as Form } from '~shared/Fields';
@@ -11,11 +9,10 @@ import { pipe, mapPayload, withMeta } from '~utils/actions';
 import { WizardDialogType } from '~hooks';
 import { formatText } from '~utils/intl';
 import { notNull } from '~utils/arrays';
-import { createAddress } from '~utils/web3';
-import { Colony, Token } from '~types';
 
 import { getTokenManagementDialogPayload } from './helpers';
 import TokenManagementDialogForm from './TokenManagementDialogForm';
+import { FormValues, getValidationSchema } from './validation';
 
 const displayName = 'common.TokenManagementDialog';
 
@@ -24,72 +21,11 @@ const MSG = defineMessages({
     id: `${displayName}.errorAddingToken`,
     defaultMessage: `Sorry, there was an error adding this token. Learn more about tokens at: https://colony.io.`,
   },
-  invalidAddress: {
-    id: `${displayName}.invalidAddress`,
-    defaultMessage: 'This is not a valid address',
-  },
-  tokenNotFound: {
-    id: `${displayName}.tokenNotFound`,
-    defaultMessage:
-      'Token data not found. Please check the token contract address.',
-  },
-  tokenIsDuplicate: {
-    id: `${displayName}.tokenIsDuplicate`,
-    defaultMessage: 'This token already exists in this colony.',
-  },
 });
 
 type Props = DialogProps &
   Partial<WizardDialogType<object>> &
   ActionDialogProps;
-
-const getValidationSchema = (colony: Colony) =>
-  object({
-    forceAction: boolean().defined(),
-    tokenAddress: string()
-      .notRequired()
-      .test(
-        'is-address',
-        () => MSG.invalidAddress,
-        (value) => !value || isAddress(value),
-      )
-      .test(
-        'is-duplicate',
-        () => MSG.tokenIsDuplicate,
-        (value) => {
-          // skip this test if value is not valid.
-          if (!value || !isAddress(value)) {
-            return true;
-          }
-
-          return !colony.tokens?.items
-            .filter(notNull)
-            .some(
-              ({ token: { tokenAddress } }) =>
-                createAddress(tokenAddress) === createAddress(value),
-            );
-        },
-      ),
-    token: object<Token>()
-      .nullable()
-      .test('doesTokenExist', '', function doesTokenExist(value, context) {
-        if (!context.parent.tokenAddress || !!value) {
-          // Skip validation if tokenAddress is empty or token has been found
-          return true;
-        }
-
-        return this.createError({
-          message: formatText(MSG.tokenNotFound),
-          path: 'tokenAddress',
-        });
-      }),
-    selectedTokenAddresses: array()
-      .of(string().address().defined())
-      .notRequired(),
-    annotationMessage: string().max(4000).notRequired(),
-  }).defined();
-
-export type FormValues = InferType<ReturnType<typeof getValidationSchema>>;
 
 const TokenManagementDialog = ({
   colony,
