+++ src/components/common/Dialogs/TokenManagementDialog/index.ts
@@ -1,1 +1,2 @@
-export { default, FormValues } from './TokenManagementDialog';
+export { default } from './TokenManagementDialog';
+export { FormValues } from './validation';
